requirements.txt
numpy
pandas
langchain>=0.1.0
faiss-cpu
sentence-transformers
nltk
langchain-community
python-dotenv
langchain_openai
streamlit
langchain_huggingface
chromadb
langchain_groq
pinecone
langchain-pinecone
youtube-transcript-api>=0.6.0
pytube>=15.0.0
openai-whisper
langchain-chroma


